use int* for carray member in IntArray (instead of long*)

+++ src/main/cpp/dedekind-mkl/IntArray.h
@@ -34,7 +34,7 @@ public:
 private:
     Context* ctx;
     jintArray jarray;
-    long* carray;
+    int* carray;
     int offset;
     bool critical;
 };
